move the status client to the api folder
