+++ src/frontend/app/api/StatusRestClient.ts
